Model width/height to Double, ComponentModel width/height renamed to gridWidth/Height.

+++ src/com/seancheey/game/Model.kt
@@ -8,6 +8,6 @@ import javafx.scene.image.Image
  */
 interface Model {
     val image: Image
-    val width: Int
-    val height: Int
+    val width: Double
+    val height: Double
 }
